Factor out onboarding/routes.ts.
