+++ frontend/lib/justfix-routes.ts
@@ -8,6 +8,10 @@ import {
   createHtmlEmailStaticPageRouteInfo,
 } from "./static-page/routes";
 import { createIssuesRouteInfo } from "./issues/routes";
+import {
+  createOnboardingRouteInfo,
+  OnboardingRouteInfo,
+} from "./onboarding/routes";
 
 /**
  * Querystring argument for specifying the URL to redirect the
@@ -78,31 +82,6 @@ function createPasswordResetRouteInfo(prefix: string) {
     verify: `${prefix}/verify`,
     confirm: `${prefix}/confirm`,
     done: `${prefix}/done`,
-  };
-}
-
-export type OnboardingRouteInfo = ReturnType<typeof createOnboardingRouteInfo>;
-
-function createOnboardingRouteInfo(prefix: string) {
-  return {
-    [ROUTE_PREFIX]: prefix,
-    latestStep: prefix,
-    step1: `${prefix}/step/1`,
-    step1AddressModal: `${prefix}/step/1/address-modal`,
-    step1ConfirmAddressModal: `${prefix}/step/1/confirm-address-modal`,
-    step3: `${prefix}/step/3`,
-    step3RentStabilizedModal: `${prefix}/step/3/rent-stabilized-modal`,
-    step3MarketRateModal: `${prefix}/step/3/market-rate-modal`,
-    step3OtherModal: `${prefix}/step/3/other-modal`,
-    step3NoLeaseModal: `${prefix}/step/3/no-lease-modal`,
-    step3LearnMoreModals: {
-      rentStabilized: `${prefix}/step/3/learn-more-rent-stabilized-modal`,
-      marketRate: `${prefix}/step/3/learn-more-market-rate-modal`,
-      noLease: `${prefix}/step/3/learn-more-no-lease-modal`,
-    },
-    step4: `${prefix}/step/4`,
-    step4TermsModal: `${prefix}/step/4/terms-modal`,
-    thanks: `${prefix}/thanks`,
   };
 }
